+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 # 1.1.0
 - bug-fixed: callback is not defined
 
+# 1.0.27-rc3
+- added: fixToken function in TiledeskService utils class
+
+# 1.0.27-rc2
+- added: specchToText function to transcript audio file
+
 # 1.0.27-rc1
 - changed: context for gpt-40 and gpt-40-mini